AI_Content_To_PDF/requirements.txt
langchain
langchain-google-genai
google-generativeai>0.8.0
python-dotenv
streamlit
requests>=2.31.0
beautifulsoup4>4.12.0
langchain-community
reportlab>=4.4.3